refactor: Remove setMapZoom
