+++ lib/actions/apiV2.js
@@ -17,7 +17,6 @@ import {
   receivedVehiclePositions,
   receivedVehiclePositionsError
 } from './api'
-import { setMapZoom } from './config'
 import { zoomToStop } from './map'
 
 /**
@@ -424,11 +423,13 @@ const fetchStopInfo = (map, stop) => {
           })
         )
 
-        dispatch(zoomToStop(map, stop))
-        // TODO: fix the zoom stuff below (add zoom param to above?)
-        if (stop?.geometries?.geoJson?.type !== 'Point') {
-          dispatch(setMapZoom(10))
-        }
+        dispatch(
+          zoomToStop(
+            map,
+            stop,
+            stop?.geometries?.geoJson?.type !== 'Point' && 10
+          )
+        )
       },
       rewritePayload: (payload) => {
         const { stop } = payload?.data
