+++ lib/actions/map.js
@@ -5,7 +5,6 @@ import getGeocoder from '@opentripplanner/geocoder'
 import { clearActiveSearch } from './form'
 import { deleteUserPlace } from './user'
 import { routingQuery } from './api'
-import { setMapCenter, setMapZoom } from './config'
 
 /* SET_LOCATION action creator. Updates a from or to location in the store
  *
@@ -50,19 +49,20 @@ export function clearLocation(payload) {
 /**
  * Centers the given map to the coordinates of the specified place.
  */
-export function setMapCenter2(map /* MapRef */, location /* Location */) {
+export function setMapCenter(map /* MapRef */, location, zoom) {
   return function () {
     const { lat, lon } = location
-    if (!isNaN(lat) && !isNaN(lon)) {
-      map.jumpTo({ center: [lon, lat] })
+    if (map && !isNaN(lat) && !isNaN(lon)) {
+      map.jumpTo({ center: [lon, lat], zoom })
     }
   }
 }
 
-export function zoomToStop(map /* MapRef */, stop) {
+export function zoomToStop(map /* MapRef */, stop, zoom) {
   return function (dispatch, getState) {
-    if (!stop || !map) return
-    map.jumpTo({ center: [stop.lon, stop.lat], zoom: 17 })
+    if (stop && map) {
+      map.jumpTo({ center: [stop.lon, stop.lat], zoom: zoom || 17 })
+    }
   }
 }
