+++ lib/actions/ui.js
@@ -13,10 +13,10 @@ import { getModesForActiveAgencyFilter, getUiUrlParams } from '../util/state'
 import { getPathFromParts } from '../util/ui'
 
 import { clearActiveSearch, parseUrlQueryString, setActiveSearch } from './form'
-import { clearLocation } from './map'
+import { clearLocation, setMapCenter } from './map'
 import { findRoute, setUrlSearch } from './api'
 import { setActiveItinerary } from './narrative'
-import { setMapCenter, setMapZoom, setRouterId } from './config'
+import { setRouterId } from './config'
 
 const updateLocale = createAction('UPDATE_LOCALE')
 
@@ -175,7 +175,7 @@ function idToParams(id, delimiter = ',') {
  * Checks URL and redirects app to appropriate content (e.g., viewed
  * route or stop).
  */
-export function matchContentToUrl(location) {
+export function matchContentToUrl(map, location) {
   // eslint-disable-next-line complexity
   return function (dispatch, getState) {
     const state = getState()
@@ -232,8 +232,7 @@ export function matchContentToUrl(location) {
         }
         console.log(lat, lon, zoom, routerId)
         // Update map location/zoom and optionally override router ID.
-        if (+lat && +lon) dispatch(setMapCenter({ lat, lon }))
-        if (+zoom) dispatch(setMapZoom({ zoom }))
+        if (+lat && +lon) dispatch(setMapCenter(map, { lat, lon }, zoom))
         // If router ID is provided, override the default routerId.
         if (routerId) dispatch(setRouterId(routerId))
         dispatch(setMainPanelContent(null))
