+++ lib/components/app/responsive-webapp.js
@@ -32,6 +32,7 @@ import MobileMain from '../mobile/main'
 import printRoutes from '../../util/webapp-print-routes'
 import webAppRoutes from '../../util/webapp-routes'
 import withLoggedInUserSupport from '../user/with-logged-in-user-support'
+import withMap from '../map/with-map'
 
 import DesktopNav from './desktop-nav'
 import PopupWrapper from './popup'
@@ -60,11 +61,10 @@ class ResponsiveWebapp extends Component {
       initZoomOnLocate,
       intl,
       location,
+      map,
       matchContentToUrl,
       query,
       setLocationToCurrent,
-      setMapCenter,
-      setMapZoom,
       title
     } = this.props
     document.title = title
@@ -96,9 +96,9 @@ class ResponsiveWebapp extends Component {
       // if in mobile mode and from field is not set, use current location as from and recenter map
       if (isMobile() && query.from === null) {
         setLocationToCurrent({ locationType: 'from' }, intl)
-        setMapCenter(pt)
+        // setMapCenter(pt)
         if (initZoomOnLocate) {
-          setMapZoom({ zoom: initZoomOnLocate })
+          // setMapZoom({ zoom: initZoomOnLocate })
         }
       }
     }
@@ -106,7 +106,7 @@ class ResponsiveWebapp extends Component {
     // main content needs to switch between, for example, a viewer and a search.
     if (!isEqual(location.pathname, prevProps.location.pathname)) {
       // console.log('url changed to', location.pathname)
-      matchContentToUrl(location)
+      matchContentToUrl(map, location)
     }
 
     // Check for change between ITINERARY and PROFILE routingTypes
@@ -134,6 +134,7 @@ class ResponsiveWebapp extends Component {
       initializeModules,
       intl,
       location,
+      map,
       matchContentToUrl,
       parseUrlQueryString,
       receivedPositionResponse,
@@ -178,7 +179,7 @@ class ResponsiveWebapp extends Component {
     // Handle routing to a specific part of the app (e.g. stop viewer) on page
     // load. (This happens prior to routing request in case special routerId is
     // set from URL.)
-    matchContentToUrl(location)
+    matchContentToUrl(map, location)
     if (location && location.search) {
       // Set search params and plan trip if routing enabled and a query exists
       // in the URL.
@@ -277,8 +278,6 @@ const mapDispatchToProps = {
   parseUrlQueryString: formActions.parseUrlQueryString,
   receivedPositionResponse: locationActions.receivedPositionResponse,
   setLocationToCurrent: mapActions.setLocationToCurrent,
-  setMapCenter: configActions.setMapCenter,
-  setMapZoom: configActions.setMapZoom,
   setPopupContent: uiActions.setPopupContent
 }
 
@@ -286,7 +285,9 @@ const history = createHashHistory()
 
 const WebappWithRouter = withRouter(
   withLoggedInUserSupport(
-    injectIntl(connect(mapStateToProps, mapDispatchToProps)(ResponsiveWebapp))
+    withMap(
+      injectIntl(connect(mapStateToProps, mapDispatchToProps)(ResponsiveWebapp))
+    )
   )
 )
