+++ lib/index.js
@@ -45,9 +45,9 @@ import BatchRoutingPanel from './components/app/batch-routing-panel'
 import BatchResultsScreen from './components/mobile/batch-results-screen'
 import BatchSearchScreen from './components/mobile/batch-search-screen'
 import FormattedMode from './components/util/formatted-mode'
-import { setAutoPlan, setMapCenter } from './actions/config'
+import { setAutoPlan } from './actions/config'
 import { getCurrentPosition } from './actions/location'
-import { setLocationToCurrent, clearLocation } from './actions/map'
+import { clearLocation, setLocationToCurrent, setMapCenter } from './actions/map'
 import { findNearbyStops } from './actions/api'
 import createCallTakerReducer from './reducers/call-taker'
 import createOtpReducer from './reducers/create-otp-reducer'
